v2

+++ libft/ft_strlcpy.c
@@ -1,4 +1,4 @@
-int ft_strlcpy(char *dst, const char *src, int size)
+size_t ft_strlcpy(char *dst, const char *src, size_t size)
 {
     int i;
